+++ app/src/main/java/eu/kanade/tachiyomi/di/AppModule.kt
@@ -36,6 +36,7 @@ import tachiyomi.domain.source.service.SourceManager
 import tachiyomi.domain.storage.service.StorageManager
 import tachiyomi.source.local.image.LocalCoverManager
 import tachiyomi.source.local.io.LocalSourceFileSystem
+import tachiyomi.source.local.metadata.LocalMangaMetadataWriter
 import uy.kohesive.injekt.api.InjektModule
 import uy.kohesive.injekt.api.InjektRegistrar
 import uy.kohesive.injekt.api.addSingleton
@@ -122,6 +123,7 @@ class AppModule(val app: Application) : InjektModule {
         addSingletonFactory { AndroidStorageFolderProvider(app) }
         addSingletonFactory { LocalSourceFileSystem(get()) }
         addSingletonFactory { LocalCoverManager(app, get()) }
+        addSingletonFactory { LocalMangaMetadataWriter(get(), get()) }
         addSingletonFactory { StorageManager(app, get()) }
 
         // Asynchronously init expensive components for a faster cold start